Mix of inline and separate compilation unit, to maximize measured performance
