+++ dali/pipeline/util/lookahead_parser.cc
@@ -1,0 +1,36 @@
+// Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "dali/pipeline/util/lookahead_parser.h"
+
+namespace dali {
+namespace detail {
+
+void LookaheadParser::SkipOut(int depth) {
+  do {
+    if (st_ == kEnteringArray || st_ == kEnteringObject) {
+      ++depth;
+    } else if (st_ == kExitingArray || st_ == kExitingObject) {
+      --depth;
+    } else if (st_ == kError) {
+      return;
+    }
+
+    ParseNext();
+  } while (depth > 0);
+}
+
+}  // namespace detail
+
+}  // namespace dali
